Change the scope in which the delete action is executed
The scope is changed for the delete action which results in getting the
user back to the mailbox when the action is done from the details screen
. The goal is to use an external scope which won't cancel the coroutines
started inside it unless the process is killed. This will ensure the
delete action not be interrupted prematurely when the user leaves the
details screen.

MAILAND-1834