Fix message disappearing when performing a move action from custom label location
The mistake was in the logic for optimistic update of message labels
after a move action. The new logic means that:
- When the action is not trash, all exclusive labels are removed from
the message, except ALL_MAIL, ALL_SENT and ALL_DRAFT, and the new label
is added.
- When the action is trash, all exclusive and non-exclusive labels are
removed from the message except ALL_MAIL, ALL_SENT and ALL_DRAFT, and
the new label is added.

MAILAND-2713