Changes the ping worker policy from replace to keep
We do not need to replace an already running worker with a new instance, as
doing so leads to cancelled workers and IO exceptions thrown from the
cancelled worker. If we keep the already running worker instead, we emit fewer
network connectivity events, some of which might be incorrect (cancelled
state will result in CONNECTED event being propagated, even if in reality
we might not be connected).

MAILAND-2741