Introduces a pending send database ID value
As the ID was previously always empty, the row would get overwritten
every time a new pending send was created. This would result in only the last
created pending send being shown correctly in the UI (as message being sent).
With this change, the id is always present and collision-free for all the
practical purposes.