Handle a scenario in which a free user is uploading contacts that have contact groups
Contact groups are not available on free accounts. When we try to upload
contacts (and groups) on free accounts an error happens because the
creation of the groups is not possible. We need to ignore this error in
ConvertLocalContactsJob and show a dialog that contact groups will not
be uploaded.

MAILAND-2811