Introduces a cleanup pending send worker
The aim of the worker is to clean up the pending sends by removing
the pending send object from the database and cancelling the enqueued
send message worker. In addition, a trigger worker is added in order
to satisfy the requirement of waiting for 2h only after being online
(we do not want to cancel the pending send if the user has been
offline the whole time). A trigger worker is used as it is the
simplest way to schedule the cleanup that will survive the app being killed.

MAILAND-2789