Fix slow actions when a folder has loaded a large number of messages
The problem was caused by a query to the contacts DB which was checking
the sender/recipient contacts for each message every time the list of
messages is emitted. The mapping of the senders/recipients to the
appropriate contacts is now done in MessageRepository by querying all
the contacts at once, which improved the performance.

MAILAND-2812