Remove the dialog that appears when more than 100 labels are selected in LabelsActionSheet
This dialog was preventing users from selecting more than 100 labels
when wanting to apply labels because we couldn't send more than 100
labels to the worker because of a size limitation regarding input
parameters. This was a problem only for conversations. However, this
was preventing users from deselecting labels as well when there were
already more than 100 selected.

We don't need this dialog anymore, because the initial issue can be
fixed by chunking input parameters.

MAILAND-2828