Fix injecting dark style in sent messages when the app is in dark mode
The problem was that we were sending a message that was styled according
to what we show in the app. Instead, for the sending, we need the clean
HTML message body of the parent message when we reply/forward. Initial
message content is the holder of the quoted text (parent message) and
the solution included setting its value to be the clean HTML body of the
parent message.

MAILAND-2847