Fix multiple "Retrieving notifications" notifications show up at the same time
The reason for this is that the id of the notification is set to be the
id of the worker, which means that for every
ProcessPushNotificationDataWorker there will be a separate notification.
This became more noticeable because of the touched flag notification.
After we do a bulk operation on messages (for example delete), a touched
flag notification is received for each of the messages. That means that
the worker is started multiple times in a row and a notification for
each instance of the worker is shown.

The solution is to make sure that all "Retrieving notifications"
notifications have the same id, so that when multiple workers are active
at the same time, there is only one notification shown.

MAILAND-2918