Sets the sentry user on the main thread
The user information from the reports has been missing till now
(with few lucky breaks here and there). Looks like the init part,
including setting the user, needs to happen on the main thread. This
is based on the mention of this in one of the github issues. A proper
question about this to the sentry team will follow.