Runs fetch/observe message list queries in transactions
This is a workaround for mailbox failing to load with:

"java.lang.IllegalStateException: Couldn't read row X, col Y from
CursorWindow.  Make sure the Cursor is initialized correctly before
accessing data from it."

This seems to happen when:
- the query is too large to fit in a single cursor page (2MB cap),
- at the same time, the underlying data changes (e.g. due to an event
coming through).

The result is inconsistent data while the cursor is switching to the
next page of results.

For more info see:
- https://issuetracker.google.com/issues/65045157
- https://issuetracker.google.com/issues/73859920