Tries to ping the main BE when the alternative has failed
This is in accordance with the outline of the DoH algorithm and how
core is handling DoH. The result of this change is that user will be
switched back to the main server more easily, if it is available,
once the alternative starts failing. This should also result in
fewer "Can't reach server" banners as with the old implementation
those could be shown even if the main BE was accessible.

MAILAND-2944