Fixes app crash when replying to a message from notification
When opening composer from the notification, the first time we try
to fetch and decrypt the parent message to which we are replying,
the parent message body is null and trying to decrypt crashes the app.
This is fixed by catching the exception in the use case and returning
a null when that happens.

MAILAND-2951