Does not load "auto:blank" when recycling a message detail view
Due to some unclear behaviour, when webView.loadUrl("about:blank") is
called, consecutive calls to load html on that same webview afterwards
result in the onPageFinished callback not being called.

We rely on that callback to show the remote content button, which is
hidden by default. If the callback is not called after loading message
html, the button will not show.

We bind a message when loading the body several times. In some
cases, recyclerview decides to recycle a viewholder, and use it
immediately afterwards to show the body. If that happens,
"auto:blank" will be loaded while recycling, after which the adapter
will attempt to load the message HTML. This attempt will succeed but
will not be followed by onPageFinished callback.

Loading "auto:blank" is the recommended way to clear a webview,
according to the docs.

In case some clearing is needed when conversations are enabled, some
alternative solutions, which should not mess with the button, include:
- disabling recycling of views on the message detail recyclerview-
should be okay if the conversations never get exceedingly long,
- instead of loading "auto:blank", load empty HTML using
loadHtmlDataIntoWebView.

CLIENT-4426