Fixes the counters flow stopping when network error happened
If an error happened while trying to fetch the counters from the BE,
the flow would stop emitting. This means the counter in the mailbox
would stop updating automatically at this point.

Fixed by catching the exception on the inner flow, otherwise, if the
exception was caught on the outer flow, the inner flow would get
cancelled and stop emitting.