Stops observing the mailbox counters when screen in bg
This is mostly a fix for excessive counter calls we make, even when
the app is in bg. Currently two calls to fetch counters are made
every 30s, even if the app is in the bg. This is quite wasteful and
puts strain on the BE (we were already made aware of the excessive
calls we make to those endpoints).