Injects the main backend api into the api manager class
In order to check if the main backend is available even when we are
switch to a proxy, we need to always have access to the main backend.
This is achieved by injecting two instances of api to the api manager
class: one that can be reset when we switch to a proxy, and the other
one always pointing to the main BE. The main BE one is only used to
execute the ping call.

MAILAND-2981