Informs the job manager as soon as the server reachable status changes
This way the jobs that require network will run as soon as we determine
that the server is indeed reachable.

Additionally, changes a name of a flow to a more meaningful one.

Finally, fixes synchronisation of the hasConn function- before it was
synchronised on a boolean, which is a mistake. On top of that, the read
and the write functions were being synchronised in different ways,
using different objects as locks.

MAILAND-2980