Does not try to switch to main BE when current legacy user is null
The ping worker would fail with an exception when the app was first
started with no user logged in. This might have consequences later
down the line as it would be used to indicate the server is not
reachable.

Assume that we are already using the main BE when the legacy user
is not present (user is not logged in).