Fix android instrumented tests failures
Caused by some ops in core's newly added initializers to happen before
the DI componenet was setup.
Based on https://github.com/google/dagger/issues/3356 this issue is
fixed in dagger / hilt 2.42. Such upgrade doesn't apply cleanly here,
which lead to this workaround of removing the offending core
initializers in the manifest to instead invoke them manually, so that
they won't be initialized when runnign tests (as they use
`HiltTestApplication` instead of `ProtonApplication`)