Disable WKD encryption when x-pm-encrypt-untrusted is set to false.
So far mail clients were supposed to encrypt by default if the server
returned untrusted encryption keys (usually retrieved from WKD).
The web client will add a feature to allow users to disable encryption
for those untrusted keys, in which case the message would be sent in
plaintext.
This is signaled by adding the 'x-pm-encrypt-untrusted: false' flag to
the signed part of the contact.

MAILAND-3015