Prevent draft body from being lost when not loaded locally
On low connectivity, opening a draft for the first time (meaning
its body was never fetched) and the exiting the composer before such
loading finished, would cause the draft to be updated with the local
content of the body - empty - causing its loss.

This commit introduces a flag that's set when the VM schedules the
"fetchDraftDetailJob" and reset when such job returns any result
(success or failure). When the user exits the composer, this flag is
checked and if found set the draft will not be saved.

CLIENT-4905