Fix issues where attachments are removed from composer
This issue's first visible cause is a race condition: if we open a new
composer, type in a recipient, subject and body and then press the "add
attachment" button _with the right timing_ the issue will happen.
The "right timing" seem to be couple of seconds: the assumption is that
the save draft flow have to be started but not finished.

This is due to the result of the "save draft" flow (which is triggered
automatically after we stop typing) is stored in a liveData which does
not get consumed by the composer activity till the "add attachments activity"
drawn on top returns. Once this live data is consumed, it overrides one
field that the "result" of add attachments activity did just set -
effectively removing the attachment.

This is the key part of this issue, as previously the livedata seemed to
be consumed anyways and thus the race condition was not encountered
(or very unlikely to happen).

While git bisect points the issue started with commit "Targeting sdk 33
to allow for granular media permissions" [6c951a] (which is the reason
for the fix being a rollback of target sdk) this doesn't seem to be the
whole story. Few other changes dependencies update happened at the
same time (kotlin update, activity and coroutines libs updated transitely
through core).

Such bits seem to be concurring to creating the issue, making it uneasy
to define one single root cause with confidence.

MAILAND-3059