Increase mailbox views count when closing a detail screen
This count will be used to define when a user can be presented with the
"rate app" dialog. We increase the count when going back from the detail
rather than at some point in the mailbox screen's lifecycle to avoid the
case where the dialog is shown when resuming mailbox from the
background.

MAILAND-3071