Let backend know that the rate app flow was triggered
by setting the feature flag to false.

This is not a guarantee that the rate app dialog was shown,
as the final decision on whether to show the "rate app dialog"
or not is done by google play lib based on "API usage quota"
(the rate dialog can only be shown so many times in a given timeframe).

MAILAND-3071