Avoid enqueueing multiple jobs to fetch updates
Due to the way we rely on AlarmManager as a cron system to schedule new jobs to
fetch updates from event loop, when offline, it might happen that
multiple instances of the same job (FetchUpdateJob) are enqueued one
after the other, resulting in duplicated API calls when back online.

Here, we cancel any pre-existing "FetchUpdateJob" job before adding a
new one to the queue.

MAILAND-3074