Make QueueNetworkUtil check for actual connectivity
As from the docs, the isConnectedOrConnecting method "is good for
applications that need to do anything related to the network other than
read or write data. For the latter, call isConnected() instead, which
guarantees that the network is fully usable.".

In the cases used by this method, we want to have a fully working network
as the usage is to transfer data.

https://developer.android.com/reference/android/net/NetworkInfo#isConnected()

MAILAND-3074